强化demo部署功能

+++ src/share/ts/utils/APIUtils.ts
@@ -114,36 +114,33 @@ export function checkUsernameAvailability(username: string) {
 }
 
 export function getServerConfig() {
-    return request({
-       url: API_BASE_URL + '/config',
-       method: 'GET'
-    });
+    if (process.env.REACT_APP_BUILD_PROFILE === 'demo') {
+        return Promise.resolve(JSON.parse(localStorage.getItem('demo_mock_server_config') || '{}'));
+    } else {
+        return request({
+           url: API_BASE_URL + '/config',
+           method: 'GET'
+        });
+    }
 }
 
 export function setServerConfig(serverConfig: ServerConfig) {
-    return request({
-        url: API_BASE_URL + '/config',
-        method: 'POST',
-        body: JSON.stringify(serverConfig)
-    });
+    if (process.env.REACT_APP_BUILD_PROFILE === 'demo') {
+        localStorage.setItem('demo_mock_server_config', JSON.stringify(serverConfig));
+        return Promise.resolve();
+    } else {
+        return request({
+            url: API_BASE_URL + '/config',
+            method: 'POST',
+            body: JSON.stringify(serverConfig)
+        });
+    }
 }
 
 export function workspaceRebuild() {
     return request({
         url: API_BASE_URL + '/config/git_refresh',
         method: 'GET'
-    });
-}
-
-export function updateDoc(docId: number, docInfo: DocInfo) {
-    // if (!localStorage.getItem(ACCESS_TOKEN) && process.env.REACT_APP_BUILD_PROFILE === 'cloud') {
-    //     return Promise.reject('No access token set.');
-    // }
-
-    return request({
-        url: API_BASE_URL + '/docs/' + docId,
-        method: 'PUT',
-        body: JSON.stringify(docInfo)
     });
 }
 
@@ -198,16 +195,42 @@ export function getSubscriptionFileContent(filepath: string) {
     });
 }
 
+export function updateDoc(docId: number, docInfo: DocInfo) {
+    if (process.env.REACT_APP_BUILD_PROFILE === 'demo') {
+        const docs = JSON.parse(localStorage.getItem('demo_mock_doc_list') || '[]') as any[];
+        localStorage.setItem('demo_mock_doc_list', JSON.stringify([{
+            content: JSON.stringify({text: ''}),
+            ...docInfo}].concat(
+              docs.filter(d => d.id !== docId)
+        )));
+        localStorage.setItem(`demo_mock_doc_content_${docId}`, docInfo.content!);
+        return Promise.resolve({message: 'save success', id: docId});
+    }
+    return request({
+        url: API_BASE_URL + '/docs/' + docId,
+        method: 'PUT',
+        body: JSON.stringify(docInfo)
+    });
+}
+
 export function uploadDoc(docInfo: DocInfo) {
     // if (!localStorage.getItem(ACCESS_TOKEN) && process.env.REACT_APP_BUILD_PROFILE === 'cloud') {
     //     return Promise.reject('No access token set.');
     // }
-
-    return request({
-        url: API_BASE_URL + '/docs/',
-        method: 'POST',
-        body: JSON.stringify(docInfo)
-    });
+    if (process.env.REACT_APP_BUILD_PROFILE === 'demo') {
+        const docs = JSON.parse(localStorage.getItem('demo_mock_doc_list') || '[]');
+        const docId = docs.length;
+        docs.push({id: docId, filename: docInfo.name + '.effect.json', ...docInfo});
+        localStorage.setItem(`demo_mock_doc_content_${docId}`, docInfo.content || JSON.stringify({text: ''}));
+        localStorage.setItem('demo_mock_doc_list', JSON.stringify(docs));
+        return Promise.resolve({message: 'save success', id: docId});
+    } else {
+        return request({
+            url: API_BASE_URL + '/docs/',
+            method: 'POST',
+            body: JSON.stringify(docInfo)
+        });
+    }
 }
 
 export function getDocVersions(docId: number) {
@@ -235,6 +258,9 @@ export function getDocContent(docId: number, version: string = 'HEAD') {
     if (docId === -1) {
         return Promise.resolve({content: config.getDefaultData()});
     }
+    if (process.env.REACT_APP_BUILD_PROFILE === 'demo') {
+        return Promise.resolve({content: localStorage.getItem(`demo_mock_doc_content_${docId}`)!});
+    }
     return request({
         url: `${API_BASE_URL}/docs/${docId}?version=${version}`,
         method: 'GET',
@@ -242,9 +268,12 @@ export function getDocContent(docId: number, version: string = 'HEAD') {
 }
 
 export function deleteDocContent(docId: number) {
-    // if (!localStorage.getItem(ACCESS_TOKEN) && process.env.REACT_APP_BUILD_PROFILE === 'cloud') {
-    //     return Promise.reject('No access token set.');
-    // }
+    if (process.env.REACT_APP_BUILD_PROFILE === 'demo') {
+        const docs = JSON.parse(localStorage.getItem('demo_mock_doc_list') || '[]') as DocInfo[];
+        localStorage.setItem('demo_mock_doc_list', JSON.stringify(docs.filter(d => d.id !== docId)));
+        localStorage.setItem(`demo_mock_doc_content_${docId}`, JSON.stringify({text: ''}));
+        return Promise.resolve();
+    }
     return request({
         url: API_BASE_URL + '/docs/' + docId,
         method: 'DELETE'
@@ -252,11 +281,12 @@ export function deleteDocContent(docId: number) {
 }
 
 export function getCurrentUserDocs() {
-    // if (!localStorage.getItem(ACCESS_TOKEN) && process.env.REACT_APP_BUILD_PROFILE === 'cloud') {
-    //     return Promise.reject('No access token set.');
-    // }
     if (process.env.REACT_APP_BUILD_PROFILE === 'demo') {
-        return Promise.resolve({content: [{name: '欢迎使用Effect笔记', filename: 'help.effect.json', tag: JSON.stringify([]), id: -1}]});
+        return Promise.resolve( {
+            content: [{name: '欢迎使用Effect笔记', filename: 'help.effect.json', tag: JSON.stringify([]), id: -1}].concat(
+              JSON.parse(localStorage.getItem('demo_mock_doc_list') || '[]')
+            )
+        });
     }
     return request({
         url: API_BASE_URL + '/docs',
